requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.0
flask==3.0.2
requests==2.31.0
pytz==2024.1
asyncpg==0.30.0
openpyxl==3.1.5